+++ docker-images/monitoring/bootstrap.sh
@@ -11,13 +11,20 @@ while [ "$(curl -m 1 -s -q -XGET 'http://localhost:8086/ping' >/dev/null || echo
 done
 
 echo "database ready"
+/usr/bin/influx -execute 'create database if not exists samsara'
 
-if [ "$(curl -s -m 1 -XGET 'http://localhost:8086/db?u=root&p=root' | grep -c samsara)" == "0" ]
-then
-    echo "creating default database"
-    curl -s -X POST 'http://localhost:8086/db?u=root&p=root' -d '{"name": "samsara"}'
+
+# wait for grafana to come up
+while [ "$(curl -m 1 -s -q -XGET 'http://admin:admin@localhost/api/org' >/dev/null || echo 1)" == "1" ] ; do
+    echo "Waiting for grafana to start up..."
+    sleep 3
+done
+
+if [ "$(curl -sS "http://$1:$!@localhost/api/datasources" | grep -q samsara && echo 1)" == "1" ] ; then
+    echo "datasource already present..."
 else
-    echo "'samsara' database already present..."
+    echo "creating datasource..."
+    curl -sS "http://$1:$2@localhost/api/datasources" -H 'Content-Type: application/json' -XPOST -d '{"name":"samsara","type":"influxdb","url":"http://127.0.0.1:8086/","access":"proxy","isDefault":true,"database":"samsara","user":"admin","password":"admin"}'
 fi
 
 exit 0